repair 102

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -1,19 +1,24 @@
 #include <stdio.h>
 
+/**
+ * main - print 50 numbers in fibonacci
+ * Return: 0
+ */
+
 int main(void)
 {
 	int i;
-	unsigned int num1 = 1;
-	unsigned int num2 = 2;
-	unsigned int sig;
+	long int num1 = 1;
+	long int num2 = 2;
+	long int sig;
 
-	printf("%u, ", num1);
-	printf("%u, ", num2);
+	printf("%ld, ", num1);
+	printf("%ld, ", num2);
 
 	for (i = 0; i < 48; i++)
 	{
 		sig = num1 + num2;
-		printf("%u", sig);
+		printf("%ld", sig);
 		if (i < 47)
 		{
 			printf(", ");
